Improve core plugins
