+++ packages/roosterjs-editor-core/lib/index.ts
@@ -29,9 +29,8 @@ export { default as Undo } from './undo/Undo';
 
 // Core Plugins
 export { default as EditPlugin } from './corePlugins/EditPlugin';
-export { default as IMEPlugin } from './corePlugins/IMEPlugin';
 export { default as MouseUpPlugin } from './corePlugins/MouseUpPlugin';
-export { default as SelectionPlugin } from './corePlugins/SelectionPlugin';
+export { default as DOMEventPlugin } from './corePlugins/DOMEventPlugin';
 export { default as TypeInContainerPlugin } from './corePlugins/TypeInContainerPlugin';
 
 // Event APIs
